몬스터 데미지 텍스트 풀링 사용해서 뜨게하기 +  스킬 원복중

+++ Assets/Scripts/Creature_HeroMonsterBase/Creature.cs
@@ -7,7 +7,7 @@ public class Creature : ScriptableObject
 {
     public string Id;
     public string Name;
-    public int MaxHp;
-    public int Atk;
+    public float MaxHp;
+    public float Atk;
     public float Speed;
 }
