+++ Assets/Scripts/Projectile/ProjectileBase.cs
@@ -7,10 +7,10 @@ using UnityEngine;
 public class ProjectileBase : SerializedMonoBehaviour
 {   
     // 스킬의 공격력을 받아올 변수
-    protected int _atk;    
+    protected float _atk;    
 
     // 스킬의 공격력을 받아옴
-    public void SetProjectileAtk(int atk)
+    public void SetProjectileAtk(float atk)
     {
         _atk = atk;
     }
